Modify to apply SVT alignment shifts.
Modified init method so it reads in the alignment shifts whenever there is a change in run number. Initial run number is -1 so it will always read in the database at least twice. This is because we don't know the run number at start-up. For run number <= 99 use all zeros for simulated data and for run number > 99 use the survey results.